Revert "FYP-11: Attempt 1"
