+++ app/(auth-pages)/forgot-password/page.tsx
@@ -31,7 +31,7 @@ export default async function ForgotPassword(props: {
           <FormMessage message={searchParams} />
         </div>
       </form>
-      {/* <SmtpMessage /> */}
+      <SmtpMessage />
     </>
   );
 }
