+++ app/(auth-pages)/sign-up/page.tsx
@@ -45,7 +45,7 @@ export default async function Signup(props: {
           <FormMessage message={searchParams} />
         </div>
       </form>
-      {/* <SmtpMessage /> */}
+      <SmtpMessage />
     </>
   );
 }
