--- app/sketch/sketch.html
@@ -1,19 +1,0 @@
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <title>Canvas Image Classification using DoodleNet and p5.js</title>
-
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.9.0/p5.min.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/p5.js/0.9.0/addons/p5.dom.min.js"></script>
-
-  <script src="https://unpkg.com/ml5@0.12.2/dist/ml5.min.js"></script>
-  <style></style>
-</head>
-
-<body>
-  <h1>Canvas Image Classification using DoodleNet and p5.js</h1>
-  <script src="sketch.js"></script>
-</body>
-
-</html>
