--- components/Navbar.tsx
@@ -1,10 +1,0 @@
-// filepath: c:\Users\MSI-PC\dyllen\fyp-doodlenet-with-supabase\components\Navbar.tsx
-import Link from "next/link";
-
-export default function Navbar() {
-  return (
-    <div className="w-full bg-primary p-4 flex items-center fixed top-0 left-0 z-10">
-      <img src="/doodle-it-logo.png" alt="Logo" className="h-8 ml-4 rounded-full" />
-    </div>
-  );
-}
