--- template/app/globals.css
@@ -1,48 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;600&display=swap');
-
-:root {
-  --primary-color: #68C0FF;
-  --secondary-color: #FFD139;
-  --accent-color: #2EAA50;
-  --text-color: #333;
-  --font-family: 'Poppins', sans-serif;
-}
-
-body {
-  font-family: var(--font-family);
-  color: var(--text-color);
-  background-color: #fff;
-}
-
-/* Button Styles */
-.btn-primary {
-  background-color: var(--secondary-color);
-  color: black;
-  font-weight: bold;
-  padding: 10px 20px;
-  border-radius: 10px;
-  transition: 0.3s;
-}
-
-.btn-primary:hover {
-  background-color: #e6b800;
-}
-
-.btn-teacher {
-  background-color: var(--accent-color);
-  color: white;
-  font-weight: bold;
-  padding: 10px 20px;
-  border-radius: 10px;
-  transition: 0.3s;
-}
-
-.btn-teacher:hover {
-  background-color: #248a40;
-}
-
-/* Header Styles */
-.header {
-  background-color: var(--primary-color);
-  padding: 10px;
-}
